Move BlockGroups.?pp to Domain/Structure
It ddidn't depend on anything in Creators and so creates a pretty strong tie to
a lot of the code. This is because Creators depends on things like H5.
