+++ tests/Unit/Domain/Creators/CMakeLists.txt
@@ -6,7 +6,6 @@ set(LIBRARY "Test_DomainCreators")
 set(LIBRARY_SOURCES
   Test_AlignedLattice.cpp
   Test_BinaryCompactObject.cpp
-  Test_BlockGroups.cpp
   Test_Brick.cpp
   Test_Cylinder.cpp
   Test_CylindricalBinaryCompactObject.cpp
